% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadGRanges.R
\name{loadGRanges}
\alias{loadGRanges}
\title{Load a GRanges}
\usage{
loadGRanges(info, project)
}
\value{
A \linkS4class{GRanges} object.
}
\description{
Load a \linkS4class{GRanges} object based on the metadata saved by the corresponding \code{\link{stageObject}} method.
}
\examples{
gr <- GRanges(c("chrA", "chrB"), IRanges(c(1, 5), c(100, 200)))
seqlengths(gr) <- c(chrA=1000, chrB=2000)

# Staging it:
tmp <- tempfile()
dir.create(tmp)
info <- stageObject(gr, tmp, path="ranges")

# Now loading it back in:
loadGRanges(info, tmp)

}
