% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadGRangesList.R
\name{loadGRangesList}
\alias{loadGRangesList}
\title{Load a GRanges}
\usage{
loadGRangesList(info, project)
}
\value{
A \linkS4class{GRangesList} object.
}
\description{
Load a \linkS4class{GRangesList} object based on the metadata saved by the corresponding \code{\link{stageObject}} method.
}
\examples{
gr <- GRanges(c("chrA", "chrB"), IRanges(c(1, 5), c(100, 200)))
seqlengths(gr) <- c(chrA=1000, chrB=2000)
grl <- split(gr, rep(1:3, length.out=length(gr)))

# Staging it:
tmp <- tempfile()
dir.create(tmp)
info <- stageObject(grl, tmp, path="ranges")

# Now loading it back in:
loadGRangesList(info, tmp)

}
