% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSeqinfo.R
\name{loadSeqinfo}
\alias{loadSeqinfo}
\title{Load a Seqinfo}
\usage{
loadSeqinfo(info, project)
}
\arguments{
\item{info}{Named list containing the metadata for this object.}

\item{project}{Any argument accepted by the acquisition functions, see \code{?\link[alabaster.base]{acquireFile}}.
By default, this should be a string containing the path to a staging directory.}
}
\value{
A \linkS4class{Seqinfo} object.
}
\description{
Load a \linkS4class{Seqinfo} object based on the metadata saved by the corresponding \code{\link{stageObject}} method.
}
\examples{
si <- Seqinfo(c("chrA", "chrB"), c(1000, 2000))

# Staging it:
tmp <- tempfile()
dir.create(tmp)
info <- stageObject(si, tmp, path="seqinfo")

# Now loading it back in:
loadSeqinfo(info, tmp)

}
