% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGRanges.R
\name{readGRanges}
\alias{readGRanges}
\alias{loadGRanges}
\title{Read a GRanges from disk}
\usage{
readGRanges(path, metadata, ...)
}
\arguments{
\item{path}{String containing a path to a directory, itself created with the \code{\link[alabaster.base]{saveObject}} method for \link[GenomicRanges]{GRanges}.}

\item{metadata}{Named list of metadata for this object, see \code{\link[alabaster.base]{readObjectFile}} for details.}

\item{...}{Further arguments to pass to internal \code{\link[alabaster.base]{altReadObject}} calls.}
}
\value{
A \link[GenomicRanges]{GRanges} object.
}
\description{
Read a \link[GenomicRanges]{GRanges} object from its on-disk representation.
This is usually not directly called by users, but is instead called by dispatch in \code{\link[alabaster.base]{readObject}}.
}
\examples{
gr <- GRanges(c("chrA", "chrB"), IRanges(c(1, 5), c(100, 200)))
seqlengths(gr) <- c(chrA=1000, chrB=2000)

tmp <- tempfile()
saveObject(gr, tmp)
readObject(tmp)

}
\seealso{
\code{"\link{saveObject,GRanges-method}"}, to save a \link[GenomicRanges]{GRanges} to disk.
}
\author{
Aaron Lun
}
