% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stageAtomicVectorList.R
\name{stageObject,CompressedAtomicList-method}
\alias{stageObject,CompressedAtomicList-method}
\title{Stage a compressed list of atomic vectors}
\usage{
\S4method{stageObject}{CompressedAtomicList}(
  x,
  dir,
  path,
  child = FALSE,
  group.name = "grouping",
  concat.name = "concatenated",
  mcols.name = "mcols",
  meta.name = "other"
)
}
\arguments{
\item{x}{A Bioconductor object of the specified class.}

\item{dir}{String containing the path to the staging directory.}

\item{path}{String containing a prefix of the relative path inside \code{dir} where \code{x} is to be saved.
The actual path used to save \code{x} may include additional components, see Details.}

\item{child}{Logical scalar indicating whether \code{x} is a child of a larger object.}

\item{group.name}{String containing the name of the file inside \code{path} to save the list groupings.}

\item{concat.name}{String containing the extension-less name of the file inside \code{path} to save the concatenated elements.}

\item{mcols.name}{String specifying the name of the directory inside \code{path} to save the \code{\link{mcols}}.
If \code{NULL}, per-element metadata is not saved.}

\item{meta.name}{String specifying the name of the directory inside \code{path} to save \code{\link{metadata}(x)}.
If \code{NULL}, object metadata is not saved.}
}
\value{
A named list containing the metadata for \code{x}.
The contents of \code{x} are saved inside \code{path} and referenced from the metadata.
}
\description{
Stage a \linkS4class{CompressedAtomicList} object.
}
\details{
The staging process will save both the interval groupings and the concatenated vector into separate files.
The concatenated vector is coerced into a DataFrame and staged with the corresponding \code{\link{stageObject}} method.
The CompressedList may also be decorated with metadata for each list element in its \code{\link{mcols}}, which is saved to \code{mcols.name}.
No file is created at \code{mcols.name} if \code{\link{mcols}(x)} is \code{NULL} or has no columns.
}
\examples{
tmp <- tempfile()
dir.create(tmp)

library(S4Vectors)
X <- splitAsList(LETTERS, sample(3, 26, replace=TRUE))
stageObject(X, tmp, path="test1")
list.files(file.path(tmp, "test1"))

}
\author{
Aaron Lun
}
