% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stageGRanges.R
\name{stageObject,GRanges-method}
\alias{stageObject,GRanges-method}
\title{Stage a GRanges object}
\usage{
\S4method{stageObject}{GRanges}(
  x,
  dir,
  path,
  child = FALSE,
  coord.name = "ranges",
  seqinfo.name = "seqinfo",
  mcols.name = "mcols",
  meta.name = "other"
)
}
\arguments{
\item{x}{A \linkS4class{GRanges} object or one of its subclasses.}

\item{dir}{String containing the path to the staging directory.}

\item{path}{String containing a prefix of the relative path inside \code{dir} where \code{x} is to be saved.
The actual path used to save \code{x} may include additional components, see Details.}

\item{child}{Logical scalar indicating whether \code{x} is a child of a larger object.}

\item{coord.name}{String containing the name of the file inside \code{path} to save the genomic coordinates.}

\item{seqinfo.name}{String containing the name of the file inside \code{path} to save the sequence information.}

\item{mcols.name}{String specifying the name of the directory inside \code{path} to save \code{\link{mcols}(x)}.
If \code{NULL}, per-element metadata is not saved.}

\item{meta.name}{String specifying the name of the directory inside \code{path} to save \code{\link{metadata}(x)}.
If \code{NULL}, object metadata is not saved.}
}
\value{
A named list containing the metadata for \code{x}.
The contents of \code{x} are saved into various files inside \code{file.path(dir, path)}.
}
\description{
Stage a \linkS4class{GRanges} object containing genomic intervals.
}
\details{
Setting \code{mcols.name=NULL} and \code{meta.name=NULL} will skip the staging of the \code{\link{mcols}} and \code{\link{metadata}}.
This is primarily useful for use in staging RangedSummarizedExperiments where the \code{\link{mcols}} have already been saved as part of the \code{rowData}.
}
\examples{
gr <- GRanges(c("chrA", "chrB"), IRanges(c(1, 5), c(100, 200)))
seqlengths(gr) <- c(chrA=1000, chrB=2000)

tmp <- tempfile()
dir.create(tmp)
stageObject(gr, tmp, path="ranges")
list.files(tmp, recursive=TRUE)

}
\author{
Aaron Lun
}
