% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stageSeqinfo.R
\name{stageObject,Seqinfo-method}
\alias{stageObject,Seqinfo-method}
\title{Stage a Seqinfo object}
\usage{
\S4method{stageObject}{Seqinfo}(x, dir, path, child = FALSE)
}
\arguments{
\item{x}{A \linkS4class{Seqinfo} object.}

\item{dir}{String containing the path to the staging directory.}

\item{path}{String containing a prefix of the relative path inside \code{dir} where \code{x} is to be saved.
The actual path used to save \code{x} may include additional components, see Details.}

\item{child}{Logical scalar indicating whether \code{x} is a child of a larger object.}
}
\value{
A named list containing the metadata for \code{x}.
The contents of \code{x} are saved into various files inside \code{file.path(dir, path)}.
}
\description{
Stage a \linkS4class{Seqinfo} object containing genomic information.
}
\examples{
si <- Seqinfo(c("chrA", "chrB"), c(1000, 2000))

tmp <- tempfile()
dir.create(tmp)
stageObject(si, tmp, path="seqinfo")
list.files(tmp, recursive=TRUE)

}
