% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAtomicVectorList.R
\name{loadAtomicVectorList}
\alias{loadAtomicVectorList}
\title{Load an atomic vector list}
\usage{
loadAtomicVectorList(info, project)
}
\arguments{
\item{info}{Named list containing the metadata for this object.}

\item{project}{Any argument accepted by the acquisition functions, see \code{?\link[alabaster.base]{acquireFile}}.
By default, this should be a string containing the path to a staging directory.}
}
\value{
A CompressedAtomicList of the relevant type.
}
\description{
Load a list of atomic vectors as a \linkS4class{CompressedAtomicList}, typically from files created by the corresponding \code{\link{stageObject}} method.
}
\examples{
library(S4Vectors)
X <- splitAsList(LETTERS, sample(3, 26, replace=TRUE))

# Staging this object:
tmp <- tempfile()
dir.create(tmp)
info <- stageObject(X, tmp, path="test1")

# Loading the object:
loadAtomicVectorList(info, tmp)

}
\author{
Aaron Lun
}
