% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataFrameList.R
\name{loadDataFrameList}
\alias{loadDataFrameList}
\title{Load a data frame list}
\usage{
loadDataFrameList(info, project)
}
\arguments{
\item{info}{Named list containing the metadata for this object.}

\item{project}{Any argument accepted by the acquisition functions, see \code{?\link[alabaster.base]{acquireFile}}.
By default, this should be a string containing the path to a staging directory.}
}
\value{
A CompressedSplitDataFrameList.
}
\description{
Load a list of data frames as a \linkS4class{CompressedSplitDataFrameList}, typically from files created by the corresponding \code{\link{stageObject}} method.
}
\examples{
Y <- splitAsList(DataFrame(Xxx=LETTERS, Yyy=1:26), sample(3, 26, replace=TRUE))

tmp <- tempfile()
dir.create(tmp)
info <- stageObject(Y, tmp, path="test2")

loadDataFrameList(info, tmp)

}
\author{
Aaron Lun
}
