% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAtomicVectorList.R
\name{readAtomicVectorList}
\alias{readAtomicVectorList}
\alias{loadAtomicVectorList}
\title{Load an atomic vector list}
\usage{
readAtomicVectorList(path, ...)
}
\arguments{
\item{path}{String containing a path to the directory, itself created with \code{\link[alabaster.base]{saveObject}} method for \linkS4class{DataFrame}s.}

\item{...}{Further arguments, passed to \code{\link[alabaster.base]{altLoadObject}} for complex nested columns.}
}
\value{
A CompressedAtomicList of the relevant type.
}
\description{
Load a list of atomic vectors as a \linkS4class{CompressedAtomicList}, typically from files created by the corresponding \code{\link{stageObject}} method.
}
\examples{
library(S4Vectors)
X <- splitAsList(LETTERS, sample(3, 26, replace=TRUE))

tmp <- tempfile()
saveObject(X, tmp)
readAtomicVectorList(tmp)

}
\author{
Aaron Lun
}
