% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGRanges.R
\name{readGRanges}
\alias{readGRanges}
\title{Load a GRanges}
\usage{
readGRanges(path, ...)
}
\arguments{
\item{path}{String containing a path to the directory, itself created with \code{\link[alabaster.base]{saveObject}} method for \linkS4class{DataFrame}s.}

\item{...}{Further arguments, passed to \code{\link[alabaster.base]{altLoadObject}} for complex nested columns.}
}
\value{
A \linkS4class{GRanges} object.
}
\description{
Load a \linkS4class{GRanges} object based on the metadata saved by the corresponding \code{\link{stageObject}} method.
}
\examples{
gr <- GRanges(c("chrA", "chrB"), IRanges(c(1, 5), c(100, 200)))
seqlengths(gr) <- c(chrA=1000, chrB=2000)

tmp <- tempfile()
saveObject(gr, tmp)
readGRanges(tmp)

}
