% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGRanges.R
\name{readGRanges}
\alias{readGRanges}
\alias{loadGRanges}
\title{Read a GRanges from disk}
\usage{
readGRanges(path, ...)
}
\arguments{
\item{path}{String containing a path to a directory, itself created with the \code{\link{stageObject}} method for \linkS4class{GRanges}.}

\item{...}{Further arguments to pass to internal \code{\link{altReadObject}} calls.}
}
\value{
A \linkS4class{GRanges} object.
}
\description{
Read a \linkS4class{GRanges} object from its on-disk representation.
}
\examples{
gr <- GRanges(c("chrA", "chrB"), IRanges(c(1, 5), c(100, 200)))
seqlengths(gr) <- c(chrA=1000, chrB=2000)

tmp <- tempfile()
saveObject(gr, tmp)
readGRanges(tmp)

}
\seealso{
\code{"\link{saveObject,GRanges-method}"}, to save a \linkS4class{GRanges} to disk.
}
